Remove "foreign code" due to ambiguity
Original:
"true if the connection was created from Luau, not C or foreign code."

"foreign code" in this case does NOT mean a connection from a foreign state but rather anything that is not Luau. Although in pure English does make sense, the general usage of the word "foreign" across the entire documentation (and especially in this specific document) is used within contexts where we are talking about things being from a different state.